Review changes made

+++ src/utils/api.js
@@ -1,13 +1,13 @@
 const baseUrl = "http://localhost:3001";
 
-export const getItems = async () => {
-  const response = await fetch(`${baseUrl}/items`);
-  const fetchedItems = await getItems();
-  console.log("Fetched items:", fetchedItems); // Log the entire list of items
-  if (!response.ok) {
-    throw new Error(`Error: ${response.status}`);
-  }
-  return response.json();
+export const getItems = () => {
+  return fetch("http://localhost:3001/items").then((res) => {
+    if (res.ok) {
+      return res.json();
+    } else {
+      return Promise.reject(`Error: ${res.status}`);
+    }
+  });
 };
 
 export const addItem = async (item) => {
@@ -26,12 +26,16 @@ export const addItem = async (item) => {
   return response.json();
 };
 
-export const deleteItem = async (id) => {
-  const response = await fetch(`${baseUrl}/items/${id}`, {
-    method: "DELETE",
-  });
-
-  if (!response.ok) {
-    throw new Error(`Error: ${response.status}`);
+export const deleteItem = (id) => {
+  if (!id) {
+    return Promise.reject("Item ID is required");
   }
+
+  return fetch(`${baseUrl}/items${id}`, {
+    method: "DELETE",
+  }).then((res) => {
+    if (!res.ok) {
+      return Promise.reject(`Error: ${res.status}`);
+    }
+  });
 };
